python/starlight/requirements.txt
snowflake-sqlalchemy==1.2.4
jinjasql==0.1.8